refactor(analytics): unify quick prompt engagement event
