+++ apps/desktop/src/renderer/src/features/analytics/reporters/agent-quick-prompt-engagement/index.ts
@@ -1,0 +1,2 @@
+export { AgentQuickPromptEngagementReporter } from "./agentQuickPromptEngagementReporter.ts";
+export type { AgentQuickPromptEngagementParams } from "./types.ts";
