+++ apps/desktop/src/renderer/src/features/analytics/reporters/agent-quick-prompt-engagement/types.ts
@@ -1,0 +1,14 @@
+import type { AgentChatEngagementBaseParams } from "../agent-chat-engagement-params.ts";
+
+export type AgentQuickPromptEngagementParams = AgentChatEngagementBaseParams &
+  (
+    | {
+        action: "panel_opened";
+        source: "composer_input";
+      }
+    | {
+        action: "prompt_used";
+        promptType: "saved" | "recommended_template";
+        source: "composer_input";
+      }
+  );
