--- apps/desktop/src/renderer/src/features/analytics/reporters/agent-quick-prompt-panel-opened/index.ts
@@ -1,2 +1,0 @@
-export { AgentQuickPromptPanelOpenedReporter } from "./agentQuickPromptPanelOpenedReporter.ts";
-export type { AgentQuickPromptPanelOpenedParams } from "./types.ts";
