--- apps/desktop/src/renderer/src/features/analytics/reporters/agent-quick-prompt-panel-opened/types.ts
@@ -1,5 +1,0 @@
-import type { AgentChatEngagementBaseParams } from "../agent-chat-engagement-params.ts";
-
-export interface AgentQuickPromptPanelOpenedParams extends AgentChatEngagementBaseParams {
-  source: "composer_input";
-}
