--- apps/desktop/src/renderer/src/features/analytics/reporters/agent-quick-prompt-used/index.ts
@@ -1,2 +1,0 @@
-export { AgentQuickPromptUsedReporter } from "./agentQuickPromptUsedReporter.ts";
-export type { AgentQuickPromptUsedParams } from "./types.ts";
