--- apps/desktop/src/renderer/src/features/analytics/reporters/agent-quick-prompt-used/types.ts
@@ -1,6 +1,0 @@
-import type { AgentChatEngagementBaseParams } from "../agent-chat-engagement-params.ts";
-
-export interface AgentQuickPromptUsedParams extends AgentChatEngagementBaseParams {
-  promptType: "saved" | "recommended_template";
-  source: "composer_input";
-}
