+++ apps/desktop/src/renderer/src/features/workspace-agent/services/internal/agentGUIEngagementAnalytics.ts
@@ -5,8 +5,7 @@ import type {
 import { AgentChatInputContentEnteredReporter } from "../../../analytics/reporters/agent-chat-input-content-entered/agentChatInputContentEnteredReporter.ts";
 import { AgentChatInputFocusedReporter } from "../../../analytics/reporters/agent-chat-input-focused/agentChatInputFocusedReporter.ts";
 import { AgentChatPanelExposedReporter } from "../../../analytics/reporters/agent-chat-panel-exposed/agentChatPanelExposedReporter.ts";
-import { AgentQuickPromptPanelOpenedReporter } from "../../../analytics/reporters/agent-quick-prompt-panel-opened/agentQuickPromptPanelOpenedReporter.ts";
-import { AgentQuickPromptUsedReporter } from "../../../analytics/reporters/agent-quick-prompt-used/agentQuickPromptUsedReporter.ts";
+import { AgentQuickPromptEngagementReporter } from "../../../analytics/reporters/agent-quick-prompt-engagement/agentQuickPromptEngagementReporter.ts";
 import type { IReporterService } from "../../../analytics/services/reporterService.interface.ts";
 import { createOptionalReporterService } from "./agentMessageSentAnalytics.ts";
 
@@ -47,15 +46,20 @@ export function createAgentGUIEngagementEventSink(input: {
         ).report();
         return;
       case "quick_prompt_panel_opened":
-        await new AgentQuickPromptPanelOpenedReporter(
-          { ...baseParams, source: event.source },
+        await new AgentQuickPromptEngagementReporter(
+          {
+            ...baseParams,
+            action: "panel_opened",
+            source: event.source
+          },
           dependencies
         ).report();
         return;
       case "quick_prompt_used":
-        await new AgentQuickPromptUsedReporter(
+        await new AgentQuickPromptEngagementReporter(
           {
             ...baseParams,
+            action: "prompt_used",
             promptType: event.promptType,
             source: event.source
           },
